backend/requirements.txt
starlette
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2
pytest-mock>=3.12.0
factory-boy>=3.3.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
authlib>=1.2.1
PyJWT>=2.8.0
cryptography>=44.0.1
loguru>=0.1.0